pyproject.toml
[project]
name = "appointmentsite"
version = "0.1.0"
description = "Online appointment booking site"
authors = [
    {name = "A.L",email = "abolfazlloukzade7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "on (>=0.0.4,<0.0.5)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "jdatetime (>=5.2.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
